Implementation of Random Forest Regression for non-linear regression model
Random Forest is a version of Ensemble Learning, it really can predict a value at the right spot, here in this implementation this model predicted the exact previous salary of an employee while fitting the model with 300 trees for this dataset. Hence it shows the more trees you use the more accurate prediction you may get for your following dataset.